Falta todo lo relacionado con el carton ademas de las salidas, y otra cosa es que no se porque no tira bien el iniciar partida con la persona que se mete en ultimo lugar yo es que no le veo sentido alguno, no creen?

+++ 3Anio/FSIV/Practica3/funciones.cpp
@@ -13,6 +13,7 @@ void help(){
 }
 
 void butterworth(Mat &filtro, float r, int n){
+
 	Mat aux = Mat(filtro.rows, filtro.cols, CV_32F);
 
 	Point centro = Point(filtro.rows/2, filtro.cols/2);
@@ -21,7 +22,7 @@ void butterworth(Mat &filtro, float r, int n){
 	for(int i=0; i<filtro.rows; i++){
 		for(int j=0; j<filtro.cols; j++){
 			radio = (double)sqrt(pow(i-centro.x, 2.0) + pow((double)(j-centro.y), 2.0));
-			aux.at<float>(i,j) = (float)(1/(1+pow((double)radio/r, 2*n)));
+			aux.at<float>(i,j) = (float)(1/(1+pow((double)(radio/r),(double) (2*n))));
 		}
 	}
 
@@ -31,23 +32,24 @@ void butterworth(Mat &filtro, float r, int n){
 }
 
 void shiftDFT(Mat &fImage){
-	Mat tmp, q0, q1, q2, q3;
+ 	
+ 	Mat tmp, q0, q1, q2, q3;
 
 	int cx = fImage.cols/2;
 	int cy = fImage.rows/2;
 
-	q0 = fImage(Rect(0, 0, cx, cy));
-	q1 = fImage(Rect(cx, 0, cx, cy));
-	q2 = fImage(Rect(0, cy, cx, cy));
-	q3 = fImage(Rect(cx, cy, cx, cy));
+		q0 = fImage(Rect(0, 0, cx, cy));
+		q1 = fImage(Rect(cx, 0, cx, cy));
+		q2 = fImage(Rect(0, cy, cx, cy));
+		q3 = fImage(Rect(cx, cy, cx, cy));
 
-	q0.copyTo(tmp);
-	q3.copyTo(q0);
-	tmp.copyTo(q3);
+		q0.copyTo(tmp);
+		q3.copyTo(q0);
+		tmp.copyTo(q3);
 
-	q1.copyTo(tmp);
-	q2.copyTo(q1);
-	tmp.copyTo(q2);
+		q1.copyTo(tmp);
+		q2.copyTo(q1);
+		tmp.copyTo(q2);
 }
 
 Mat create_spectrum(Mat &complexImg){
